--- src/app/naturestory/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export default function ColumnPage() {
-  return (
-    <div className="w-screen mt-40 flex justify-center items-center">
-      <h2>준비중입니다....</h2>
-    </div>
-  );
-}
