+++ src/components/nature-story/SideBar.tsx
@@ -1,23 +1,34 @@
-'use client';
 import Link from 'next/link';
-import { usePathname } from 'next/navigation';
 
-export function SideBar() {
-  const params = usePathname();
+import type { Props } from '@/types/page-props.type';
+
+export function SideBar({ params }: Props) {
   return (
     <nav className="flex flex-col justify-start items-center w-[180px]">
-      <Link href="/naturestory" className="px-6 py-4">
-        <h5 className={`${params === '/naturestory' ? '' : 'text-lg font-medium opacity-50'}`}>전체보기</h5>
-      </Link>
-      <Link href="/naturestory/youtube" className="px-6 py-4">
-        <h5 className={`${params === '/naturestory/youtube' ? '' : 'text-lg font-medium opacity-50'}`}>유튜브</h5>
-      </Link>
-      <Link href="/naturestory/column" className="px-6 py-4">
-        <h5 className={`${params === '/naturestory/column' ? '' : 'text-lg font-medium opacity-50'}`}>칼럼</h5>
-      </Link>
-      <Link href="/naturestory/news" className="px-6 py-4">
-        <h5 className={`${params === '/naturestory/news' ? '' : 'text-lg font-medium opacity-50'}`}>뉴스</h5>
-      </Link>
+      {categories.map(item => (
+        <Link href={`nature-story/${item.path}`} key={item.title} className="px-6 py-4">
+          <h5 className={`${params.slug === item.path ? '' : 'text-lg font-medium opacity-50'}`}>{item.title}</h5>
+        </Link>
+      ))}
     </nav>
   );
 }
+
+const categories = [
+  {
+    title: '전체보기',
+    path: '',
+  },
+  {
+    title: '유튜브',
+    path: 'youtube',
+  },
+  {
+    title: '칼럼',
+    path: 'column',
+  },
+  {
+    title: '뉴스',
+    path: 'news',
+  },
+];
