+++ src/components/nature-story/index.ts
@@ -1,3 +1,2 @@
 export * from './Content';
-export * from './NaturePage';
 export * from './SideBar';
